feat(FA-51): Add functionality to like artwork.
This commit adds the functionality to like artwork. When the user clicks the "Like" button, an API request is sent to the server to register the like for the corresponding artwork. If the request is successful, the "liked" class is added to the like button, providing visual feedback to the user. If the request fails, an error message is displayed. This feature enhances user engagement and interaction with the artwork.