feat(FA-46): Implement artwork update functionality
This commit adds the functionality to update artwork information on the
artwork edit page. When the user submits the form, an AJAX request is
made to the server with the updated artwork data. The server handles the
update operation and returns a success message if the update is
successful. If there is an error, an error message is displayed. Upon
successful update, the user is redirected to the artwork detail page for
the updated artwork.